requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
docstring_parser==0.16
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpc-google-iam-v1==0.14.1
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
numpy==2.2.3
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
shapely==2.0.7
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vertexai==1.71.1
